pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.0",
    "openai>=1.77.0",
    "python-dotenv>=1.1.0",
]
